Move GDA-related functions
