+++ docs/api.rst
@@ -126,11 +126,11 @@ GDA integration tools
 
 **Read geometry and detector parameters from GDA-generated JSON files**
 
-.. autofunction:: nexgen.tools.GDAjson2params.read_geometry_from_json
+.. autofunction:: nexgen.beamlines.GDAtools.GDAjson2params.read_geometry_from_json
 
-.. autofunction:: nexgen.tools.GDAjson2params.read_detector_params_from_json
+.. autofunction:: nexgen.beamlines.GDAtools.GDAjson2params.read_detector_params_from_json
 
 **Gather beamline and collection information from GDA-generated xml file**
 
-.. autoclass:: nexgen.tools.ExtendedRequest.ExtendedRequestIO
+.. autoclass:: nexgen.beamlines.GDAtools.ExtendedRequest.ExtendedRequestIO
     :members:
